backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
